requirements.txt
#pip install -r requirements.txt
opencv-python
numpy
Pillow
ultralytics
mediapipe